# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AnimeVault API",
        "description": "OpenAPI specification for the AnimeVault backend",
        "contact": {
            "email": "support@animevault.local"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/anime": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Lists anime with pagination, filters and sorting",
                "operationId": "apiAnimeIndex",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Searches by anime ID or title (romaji, english, native)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "maxLength": 255
                        },
                        "example": "cowboy bebop"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filters by one or more status codes. Use comma-separated values.",
                        "required": false,
                        "style": "form",
                        "explode": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "example": [
                            "FINISHED",
                            "RELEASING"
                        ]
                    },
                    {
                        "name": "format",
                        "in": "query",
                        "description": "Filters by one or more format codes. Use comma-separated values.",
                        "required": false,
                        "style": "form",
                        "explode": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "example": [
                            "TV",
                            "MOVIE"
                        ]
                    },
                    {
                        "name": "season",
                        "in": "query",
                        "description": "Filters by one or more season codes. Use comma-separated values.",
                        "required": false,
                        "style": "form",
                        "explode": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "example": [
                            "SPRING",
                            "FALL"
                        ]
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filters by one or more source codes. Use comma-separated values.",
                        "required": false,
                        "style": "form",
                        "explode": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "example": [
                            "MANGA",
                            "ORIGINAL"
                        ]
                    },
                    {
                        "name": "genres",
                        "in": "query",
                        "description": "Filters by one or more genres. Use comma-separated values.",
                        "required": false,
                        "style": "form",
                        "explode": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        },
                        "example": [
                            "Action",
                            "Drama"
                        ]
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "description": "Filters by season year",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 2100,
                            "minimum": 1900
                        },
                        "example": 2023
                    },
                    {
                        "name": "is_adult",
                        "in": "query",
                        "description": "Filters adult content",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        },
                        "example": false
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sort order for the results",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "popularity_desc",
                            "enum": [
                                "popularity_desc",
                                "score_desc",
                                "favourites_desc",
                                "recently_updated",
                                "start_date_desc",
                                "title_asc"
                            ]
                        },
                        "example": "popularity_desc"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15,
                            "maximum": 50,
                            "minimum": 1
                        },
                        "example": 15
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1,
                            "minimum": 1
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Paginated anime list"
                    }
                }
            }
        },
...
        "/api/v1/anime/{id}": {
            "get": {
                "tags": [
                    "Anime"
                ],
                "summary": "Gets a single anime with its related metadata",
                "operationId": "apiAnimeShow",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Anime identifier",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Anime found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "titles": {
                                            "properties": {
                                                "romaji": {
                                                    "type": "string",
                                                    "example": "Cowboy Bebop",
                                                    "nullable": true
                                                },
                                                "english": {
                                                    "type": "string",
                                                    "example": "Cowboy Bebop",
                                                    "nullable": true
                                                },
                                                "native": {
                                                    "type": "string",
                                                    "example": "カウボーイビバップ",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "genres": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Anime not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Anime not found."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumBearerAuth": {
                "type": "http",
                "description": "Bearer token returned by the Sanctum login or register endpoints",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}